Major refactor to vehicle spawner and menu
VehicleMenu changes
* Add bulletproof tires, and vehicle armor level to menu.
* Label vehicle menu quite a bit more
* Movel vehicle hashes to near the top of the functions.
* Move menu titles to the top of the functions.
* Remove debug strings, remove extra code.
* Add SUV sub menu.

MiscScripts changes
* Add Model namespace
* Add Request to model namespace, for requesting models.
* Add MarkAsNoLongerNeeded to model namespace.
* Add IsInCdImage, and IsAVehicle functions to model namespace.

VehicleSpawner changes
* Make vehicle spawner remove old vehicle if a new one is spawned in.
* Make spawner use my new model request function.
* Make vehicle variable accessable in the VehicleSpawner class with 'vehicleToSpawn' variable.

VehicleScripts changes
* Add SetArmor function.
* Add boilerplate for setting some vehicle upgrades, I need to set these up.
* Make RepairTires function for repairing the current vehicle tires.
* Add Enable/Disable bullet proof functions for the tires.

Other changes
* Make most console logging items use my new log_output function instead of std::cout.
* Add kcTrainerString, and kcErrorString to defines.h
* Add log_error to defines.h, log to console for errors.
* Remove VEHICLE_SCRIPTS_SINGLETON preprocessor
* Remove some extra DISABLED_CODE preprocessors in GTAped.cpp.

* Label KCMenuScript a bit more.
* Add SUV vehicle category sub menu to KCMainMenu.
* Move cops kiled display out of the debug menu.

* Fix a couple of warnings.

Added reset toggles
* Add Reset cops killed menu position, and stat class test in Stats.h.

+++ KCTrainerV-menu/Scripts/PedScripts.cpp
@@ -440,7 +440,9 @@ void PedScripts::GivePedRandomWeapons(Ped ped)
 {
 	const auto& weapons = Memory::GetAllWeapons();
 	
-	Hash wep = weapons[g_Random.GetRandomInt(0, weapons.size() - 1)];
+	//Hash wep = weapons[g_Random.GetRandomInt(0, weapons.size() - 1)];
+	// TODO Test this, should fix this warning.
+	Hash wep = weapons[g_Random.GetRandomInt(0, static_cast<int>(weapons.size()) - 1)];
 
 	GIVE_WEAPON_TO_PED(ped, wep, 9999, true, true);
 	SET_CURRENT_PED_WEAPON(ped, wep, true);
@@ -455,7 +457,9 @@ void PedScripts::GiveAllPedsRandomWeapons()
 	const auto& weapons = Memory::GetAllWeapons();
 	for (Ped ped : GetAllPeds())
 	{
-		Hash wep = weapons[g_Random.GetRandomInt(0, weapons.size() - 1)];
+		//Hash wep = weapons[g_Random.GetRandomInt(0, weapons.size() - 1)];
+		// TODO Test this, should fix this warning.
+		Hash wep = weapons[g_Random.GetRandomInt(0, static_cast<int>(weapons.size()) - 1)];
 
 		GIVE_WEAPON_TO_PED(ped, wep, 9999, true, true);
 		SET_CURRENT_PED_WEAPON(ped, wep, true);
